refactor to seaborn_file
