+++ setup.py
@@ -1,5 +1,6 @@
-from setuptools import setup, find_packages
+from setuptools import setup
 import os
+
 with open(os.path.join(os.path.dirname(__file__), 'README.rst')) as f:
     long_description = f.read()
 
@@ -18,8 +19,7 @@ setup(
     install_requires=[
     ],
     extras_require={},
-    packages=['seaborn'] + ['seaborn.' + i
-                            for i in find_packages(where='./seaborn')],
+    packages=['seaborn_file'],
     license='MIT License',
     classifiers=[
         'Intended Audience :: Developers',
